Almost complete rework, after going through a bootcamp. (#2)
reworked function-defining macro (tailor)

implemented kunge-kutta integration

added R-K of 7th order and corresponding tests

removed Option from `Reversible` trait -- NaN is a correct way to indicate that the value is not valid. Other types presumable should have these as well.

+++ gen_math_lib/src/numbers.rs
@@ -4,7 +4,7 @@ pub fn pochhammer(n: &u128, k: &u128) -> Option<u128> {
         if result > u128::MAX / p {
             return None;
         }
-        result *= p as u128;
+        result *= p;
     }
     Some(result)
 }
